Navigate to and Interact with an Element Not on the Page
"Project Y" is a piece of code that I plan to use in the construction of an e-commerce project that I have named. It is a web test automation written in Python Selenium. This piece of code allows us to find and interact with a web element that is not visible when the website is open.